+++ src/savefile.h
@@ -89,6 +89,7 @@ int rd_messages(void);
 int rd_monster_memory(void);
 int rd_object_memory(void);
 int rd_quests(void);
+int rd_quests_1(void);
 int rd_artifacts(void);
 int rd_player(void);
 int rd_ignore(void);